+++ src/core/components/InfoDialog/styles.scss
